new class GeneratorPdf
